+++ src/view/no-event-view.js
@@ -1,23 +1,8 @@
-import { createElement } from '../utils';
+import AbstractView from './abstract-view';
 
 const createNoEventTemplate = () => '<p class="trip-events__msg">Click New Event to create your first point</p>';
-
-export default class NoEventView {
-  #element = null;
-
-  get element() {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class NoEventView extends AbstractView {
   get template() {
     return createNoEventTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
