+++ src/view/trip-filter-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils';
+import AbstractView from './abstract-view';
 
 const createTripFilterTemplate = () => (
   `<form class="trip-filters" action="#" method="get">
@@ -21,22 +21,9 @@ const createTripFilterTemplate = () => (
   </form>`
 );
 
-export default class TripFilterView {
-  #element = null;
-
-  get element() {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class TripFilterView extends AbstractView {
   get template() {
     return createTripFilterTemplate();
   }
 
-  removeElement() {
-    this.#element = null;
-  }
 }
